Store the upload target as a field of the uppy upload record
There were a couple of issues with the original implementation:

The browser was reusing the same File object instance when
uploading the same file multiple times. Since that object was
used as the key in the WeakMap, we couldn't properly track
multiple upload targets for the same file.

Since that File object was never being garbage collected, the
WeakMap never lost any entries, defeating the purpose of it
being a WeakMap.